+++ examples/Makefile-posix
@@ -107,6 +107,7 @@ configure_OPTIONS                   = \
     --enable-jam-detection            \
     --enable-joiner                   \
     --enable-legacy                   \
+    --enable-mac-filter               \
     --enable-mtd-network-diagnostic   \
     --enable-raw-link-api             \
     --enable-tmf-proxy                \
