--- src/core/mac/mac_filter_stub.hpp
@@ -1,75 +1,0 @@
-/*
- *  Copyright (c) 2017, The OpenThread Authors.
- *  All rights reserved.
- *
- *  Redistribution and use in source and binary forms, with or without
- *  modification, are permitted provided that the following conditions are met:
- *  1. Redistributions of source code must retain the above copyright
- *     notice, this list of conditions and the following disclaimer.
- *  2. Redistributions in binary form must reproduce the above copyright
- *     notice, this list of conditions and the following disclaimer in the
- *     documentation and/or other materials provided with the distribution.
- *  3. Neither the name of the copyright holder nor the
- *     names of its contributors may be used to endorse or promote products
- *     derived from this software without specific prior written permission.
- *
- *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- *  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- *  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- *  ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
- *  LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- *  CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- *  SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- *  INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- *  CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- *  ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- *  POSSIBILITY OF SUCH DAMAGE.
- */
-
-/**
- * @file
- *   This file includes definitions for IEEE 802.15.4 frame filtering based on MAC address.
- */
-
-#ifndef MAC_FILTER_HPP_
-#define MAC_FILTER_HPP_
-
-#include "utils/wrap_stdint.h"
-
-#include <openthread/types.h>
-
-#include "mac/mac_frame.hpp"
-
-namespace ot {
-namespace Mac {
-
-class Filter
-{
-public:
-    typedef otMacFilterEntry Entry;
-    Filter(void) { }
-    uint8_t GetMaxEntries(void) const { return 0; }
-    otMacFilterAddressMode GetAddressMode(void) const { return OT_MAC_FILTER_ADDRESS_MODE_DISABLED; }
-    otError SetAddressMode(otMacFilterAddressMode) { return OT_ERROR_NOT_IMPLEMENTED; }
-
-    otError AddAddress(const ExtAddress &aExtAddress) { return OT_ERROR_NOT_IMPLEMENTED; }
-    otError AddAddressRssIn(const ExtAddress &aExtAddress, int8_t) { return OT_ERROR_NOT_IMPLEMENTED; }
-    otError RemoveAddress(const ExtAddress &)  { return OT_ERROR_NOT_IMPLEMENTED; }
-    void ClearAddresses(void) { }
-    otError GetNextAddress(otMacFilterIterator &, Entry &) { return OT_ERROR_NOT_IMPLEMENTED; }
-    otError AddRssIn(const ExtAddress *, int8_t) { return OT_ERROR_NOT_IMPLEMENTED; }
-    otError RemoveRssIn(const ExtAddress *) { return OT_ERROR_NOT_IMPLEMENTED; }
-    void ClearRssIn(void) { }
-    otError GetNextRssIn(otMacFilterIterator &, Entry &) { return OT_ERROR_NOT_IMPLEMENTED; }
-    otError Apply(const ExtAddress &, int8_t &) { return OT_ERROR_NONE; }
-};
-
-/**
- * @}
- *
- */
-
-}  // namespace Mac
-}  // namespace ot
-
-#endif  // MAC_FILTER_HPP_
